Validation completed. AJAX mostly completed.

+++ WhoteachswhatJPA/src/main/webapp/views_ajax_requests/viewCourseInSemesterAJAX.js
@@ -28,8 +28,9 @@
 
 	   if (!courseId|| !courseId.length) { // If no id in query string, use the one from the form
 			courseId = document.getElementById("courseId").value;
-		}
+	   }
 	   
+	   alert(courseId);
 		
 	   	for (var i = 0, j = 0; i < addForm.length - 2; ++i) {
 			if (validate_empty(addForm.elements[i].value)) {
@@ -54,7 +55,7 @@
 			elementsId[j++] = semesterId.getAttribute("id");
 		}
 		
-		if (!validate_integer(courseId.value)) {
+		if (!validate_integer(courseId)) {
 			valid = false;
 			errors[j] = document.getElementById("courseId").getAttribute("name") + "  must be a number";
 			elementsId[j++] = document.getElementById("courseId").getAttribute("id");
@@ -65,13 +66,6 @@
 				type: "POST",
 				url: "api/courseinsemester",
 				data: { 
-					addition_attribute: document.getElementById("addition_attribute").value,
-		  	   		eval_1_ans: document.getElementById("eval_1_ans").value,
-		  	   		eval_2_ans: document.getElementById("eval_2_ans").value,
-		  	 		eval_3_ans: document.getElementById("eval_3_ans").value,
-		  	 		eval_1: document.getElementById("eval_1").value,
-		  	 		eval_2: document.getElementById("eval_2").value,
-		  	 		eval_3: document.getElementById("eval_3").value,
 		  	 		sectionNumber: sectionNumber.value,
 		  	 		year: year.value,
 		  	 		course_id: courseId,
@@ -85,7 +79,7 @@
 							title : 'New Course Section in Semester added',
 							type : 'info',
 							text : 'Course has been allocated ' + document.getElementById("sectionNumber").value 
-								+ 'number of sections'
+								+ ' number of sections'
 						});
 			    		
 			    		// Form needs resetting due to never being submitted
@@ -116,7 +110,7 @@
 	
 	var updateCourseInSemester=function() {
 		var valid = true;
-		var addForm = document.getElementById("addCourseInSemesterForm");
+		var updateForm = document.getElementById("updateCourseInSemesterForm");
 		var errors = new Array();
 		var elementsId = new Array();
 	   
@@ -125,25 +119,27 @@
 		var sectionNumber = document.getElementById("sectionNumber");
 		var semesterId = document.getElementById("semesterId");
 		var year = document.getElementById("year");
+		
+		var cisId = document.getElementById("cisId");
 
 		if (!courseId|| !courseId.length) { // If no id in query string, use the one from the form
 			courseId = document.getElementById("up_courseId").value;
 		}
 	   
 		
-		for (var i = 0, j = 0; i < addForm.length - 2; ++i) {
-	   		if (validate_empty(addForm.elements[i].value)) {
+		for (var i = 0, j = 0; i < updateForm.length - 2; ++i) {
+	   		if (validate_empty(updateForm.elements[i].value)) {
 				valid = false;
-				errors[j] = addForm.elements[i].getAttribute("name") + " is required";
+				errors[j] = updateForm.elements[i].getAttribute("name") + " is required";
 				elementsId[j++] = addForm.elements[i].getAttribute("id");
 			}
 			else {
-				if (!idPattern.test(addForm.elements[i].getAttribute("id")) 
-						&& !validate_float(addForm.elements[i].value, floatLength)) {
+				if (!idPattern.test(updateForm.elements[i].getAttribute("id")) 
+						&& !validate_float(updateForm.elements[i].value, floatLength)) {
 					valid = false;
-				   	errors[j] = addForm.elements[i].getAttribute("name") + " must be " +
+				   	errors[j] = updateForm.elements[i].getAttribute("name") + " must be " +
 				   		"a number less than " + floatLength + " digits long";
-				   	elementsId[j++] = addForm.elements[i].getAttribute("id");
+				   	elementsId[j++] = updateForm.elements[i].getAttribute("id");
 				}
 			}
 	   	}
@@ -154,28 +150,27 @@
 			elementsId[j++] = semesterId.getAttribute("id");
 		}
 		
-		if (!validate_integer(courseId.value)) {
+		if (!validate_integer(courseId)) {
 			valid = false;
 			errors[j] = document.getElementById("up_courseId").getAttribute("name") + "  must be a number";
 			elementsId[j++] = document.getElementById("up_courseId").getAttribute("id");
+		}
+		
+		if (!validate_integer(cisId.value)) {
+			valid = false;
+			errors[j] = cisId.getAttribute("name") + "  must be a number";
+			elementsId[j++] = cisId.getAttribute("id");
 		}
 	   
 	   if (valid) {
 			$.ajax({
 				type: "POST",
-				url: "api/courseinsemester/"+document.getElementById("cisId").value,
+				url: "api/courseinsemester/"+cisId.value,
 				data: { 
-					addition_attribute: document.getElementById("up_addition_attribute").value,
-		  	   		eval_1_ans: document.getElementById("up_eval_1_ans").value,
-		  	   		eval_2_ans: document.getElementById("up_eval_2_ans").value,
-		  	 		eval_3_ans: document.getElementById("up_eval_3_ans").value,
-		  	 		eval_1: document.getElementById("up_eval_1").value,
-		  	 		eval_2: document.getElementById("up_eval_2").value,
-		  	 		eval_3: document.getElementById("up_eval_3").value,
-		  	 		sectionNumber: document.getElementById("up_sectionNumber").value,
-		  	 		year: document.getElementById("up_year").value,
+		  	 		sectionNumber: sectionNumber.value,
+		  	 		year: year.value,
 		  	 		course_id: courseId,
-		  	 		semester_id: document.getElementById("up_semester_id").value
+		  	 		semester_id: semesterId.value
 			   	},
 				dataType: "json",
 				cache: false,
@@ -213,23 +208,43 @@
 	   }
 	};
    
-  	var deleteCourseInSemester= function(id, course_code, semester, year) {
- 		$.ajax({
- 			type:"DELETE", 
-  			url : "api/courseinsemester/"+id,
-  			data : null,
-  			cache : false,
-  			success : function(data){
-     			if (data.success === "true") {
-    				$.pnotify({
-						title : 'Course :' + course_code + 'sections removed',
-						type : 'info',
-						text : 'Course sections has been removed for ' + semester + ' ' + year
-					});
-    		//	location.reload();
-  				}
-  	 		 }
- 		});
+  	var deleteCourseInSemester= function() {
+  		var valid = true;
+		var errors = "";
+		
+		var cisId = document.getElementById('del_cisId');
+		
+		if (!validate_integer(cisId.value)) {
+			valid = false;
+			errors = cisId.getAttribute("name") + "  must be a number";
+		}
+		
+		if(valid) {
+	 		$.ajax({
+	 			type:"DELETE", 
+	  			url : "api/courseinsemester/"+cisId.value,
+	  			data : null,
+	  			cache : false,
+	  			success : function(data){
+	     			if (data.success === "true") {
+	    				$.pnotify({
+							title : 'Course: ' + document.getElementById("del_courseCode").value + ' sections removed',
+							type : 'info',
+							text : 'Course sections has been removed for ' + document.getElementById("del_semesterName").value
+								+ ' ' + document.getElementById("del_year").value
+						});
+	    		//	location.reload();
+	  				}
+	  	 		 }
+	 		});
+		}
+ 		else { // Handle HTML inspect errors
+			$.pnotify({
+				title : 'Error',
+				type : 'info',
+				text : errors
+			});
+		}
 	};	  	   
 
 	var updateForm=function(cisId, up_sectionNumber, up_year, up_course_id, up_semester_id){
@@ -238,4 +253,11 @@
 		$("#up_year").val(up_year);
 		$("#up_courseId").val(up_course_id);
 		$("#up_semesterId").val(up_semester_id);
+	};
+	
+	var deleteForm=function(del_cisId, del_courseCode, del_semesterName, del_year) {
+		$("#del_cisId").val(del_cisId);
+		$("#del_courseCode").val(del_courseCode);
+		$("#del_semesterName").val(del_semesterName);
+		$("del_year").val(del_year);
 	};
